# docs/swagger.yaml
basePath: /api/v1
definitions:
  CreateAccountInput:
    properties:
      currency:
        type: string
      owner:
        type: string
    required:
    - currency
    - owner
    type: object
  ListAccountRequest:
    properties:
      pageID:
        minimum: 1
        type: integer
      pageSize:
        maximum: 10
        minimum: 5
        type: integer
    required:
    - pageID
    - pageSize
    type: object
  models.Account:
    properties:
      currency:
        type: string
      id:
        type: integer
      owner:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: rahul.r@i-exceed.com
    name: Iexceed technology solutions
    url: https://www.i-exceed.com/contact-us/
  description: A rest based service in Go using Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://tos.iexceed.dev
  title: Fund transfer service
  version: "1.0"
paths:
  /accounts:
    get:
      description: Responds with the list of all accounts as JSON.
      parameters:
      - description: Provide the pageId from where the records needs to be returned
        in: query
        name: page_id
        required: true
        type: integer
      - description: provide the size of the page
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListAccountRequest'
        "400":
          description: Bad/Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get Accounts based on pageId and size
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Takes a account JSON and store in DB. Return saved JSON.
      parameters:
      - description: Account JSON
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/CreateAccountInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CreateAccountInput'
        "400":
          description: Bad/Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      description: Delete an account with the given id
      parameters:
      - description: delete account by id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad/Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete account by id
      tags:
      - accounts
    get:
      description: Returns the account whose id value matches the isbn.
      parameters:
      - description: search account by id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad/Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get single account by id
      tags:
      - accounts
    put:
      description: Update an account with the given id
      parameters:
      - description: update account by id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad/Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Update account by id
      tags:
      - accounts
swagger: "2.0"